% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Communities.R
\docType{data}
\name{Communities}
\alias{Communities}
\title{Data for Communities within and nearby La Tigra National Park, Honduras}
\format{
  Simple feature collection with 104 features and 8 fields:
  \itemize{
    \item{COD_CASERI} {}
    \item{CASERI} {}
    \item{COOR_X} {}
    \item{COOR_Y} {}
    \item{COD_MUNI} {}
    \item{COD_ALDEA} {}
    \item{POB_TOTAL} {}
    \item{UbiPNLT} {}
  }
}
\source{
  Instituto Nacional de Estadística, Censo Poblacional de la República de Honduras year 2001
}
\usage{
  Communities
}
\description{
  Data for Communities within and nearby La Tigra National Park, Honduras
}
\examples{
  if (requireNamespace("sf", quietly = TRUE)) {
    library(sf)
    data(Communities)
    plot(st_geometry(Communities), axes=TRUE)
  }
}
\references{
  \url{http://170.238.108.227/binhnd/RpWebEngine.exe/Portal?BASE=CPVHND2001&lang=ESP}
}
\keyword{datasets}
\keyword{sf}
