% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Departments.R
\docType{data}
\name{Departments}
\alias{Departments}
\title{Data for Departments within and nearby La Tigra National Park, Honduras}
\format{
  Simple feature collection with 1 features and 6 fields:
    \itemize{
      \item{ID} {}
      \item{DEPARTAMEN} {}
      \item{AREA} {}
      \item{PERIMETER} {}
      \item{HECTARES} {}
      \item{COD_DEPTO} {}
    }
}
\source{
  Instituto Nacional de Estadística, Censo Poblacional de la República de Honduras year 2001
}
\usage{
  Departments
}
\description{
  Data for Departments within and nearby La Tigra National Park, Honduras
}
\examples{
  if (requireNamespace("sf", quietly = TRUE)) {
    library(sf)
    data(Departments)
    plot(st_geometry(Departments), axes=TRUE)
  }
}
\references{
  \url{http://170.238.108.227/binhnd/RpWebEngine.exe/Portal?BASE=CPVHND2001&lang=ESP}
}
\keyword{datasets}
\keyword{sf}
