\name{pnlt}
\alias{pnlt}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
  Spatial Data package of La Tigra National Park, Honduras
}
\description{
Acceses spatial data from La Tigra National Park, Honduras}
\usage{
  downloadLaTigra()
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{No Arguments Needed} {}
}

\value{Download data to local computer in format gpkg}

\author{Klaus Wolfgang Wiese Acosta}

\examples{
downloadLaTigra()
}

