% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fires1998.R
\docType{data}
\name{Fires1998}
\alias{Fires1998}
\title{Data for Fires for year 1998 from La Tigra National Park, Honduras}
\format{
  Simple feature collection with 29 features and 2 fields
    \itemize{
      \item{id} {}
      \item{Area_ha} {}
    }
}
\source{
  fire area estimation using landsat images from La Tigra National Park for the year 1998.
}
\usage{
  Fires1998
}
\description{
  Data for Fires for La Tigra National Park, Honduras for the year 1998
}
\examples{
  if (requireNamespace("sf", quietly = TRUE)) {
    library(sf)
    data(Fires1998)
    plot(st_geometry(Fires1998), axes=TRUE, col="darkred")
  }
}
\references{
  \url{https://www.amitigra.com/}
}
\keyword{datasets}
\keyword{sf}
