% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fires2000.R
\docType{data}
\name{Fires2000}
\alias{Fires2000}
\title{Data for Fires for year 2000 from La Tigra National Park, Honduras}
\format{
  Simple feature collection with 20 features and 2 fields
  \itemize{
    \item{id} {}
    \item{Area_ha} {}
  }
}
\source{
  fire area estimation using landsat images from La Tigra National Park for the year 2000.
}
\usage{
  Fires2000
}
\description{
  Data for Fires for La Tigra National Park, Honduras for the year 2000
}
\examples{
  if (requireNamespace("sf", quietly = TRUE)) {
    library(sf)
    data(Fires2000)
    plot(st_geometry(Fires2000), axes=TRUE, col="darkred")
  }
}
\references{
  \url{https://www.amitigra.com/}
}
\keyword{datasets}
\keyword{sf}
