% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fires2001.R
\docType{data}
\name{Fires2001}
\alias{Fires2001}
\title{Data for Fires for year 2001 from La Tigra National Park, Honduras}
\format{
  Simple feature collection with 37 features and 2 fields
  \itemize{
    \item{id} {}
    \item{Area_ha} {}
  }
}
\source{
  fire area estimation using landsat images from La Tigra National Park for the year 2001.
}
\usage{
  Fires2001
}
\description{
  Data for Fires for La Tigra National Park, Honduras for the year 2001
}
\examples{
  if (requireNamespace("sf", quietly = TRUE)) {
    library(sf)
    data(Fires2001)
    plot(st_geometry(Fires2001), axes=TRUE, col="darkred")
  }
}
\references{
  \url{https://www.amitigra.com/}
}
\keyword{datasets}
\keyword{sf}
