% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hills.R
\docType{data}
\name{Hills}
\alias{Hills}
\title{Data names of Hills in La Tigra National Park, Honduras}
\format{
  Simple feature collection with 47 features and 3 fields:
    \itemize{
      \item{id} {}
      \item{name} {}
      \item{elevation} {}
    }
}
\source{
  Instituto Geográfico Nacional, escala 1:50,000.
}
\usage{
  Hills
}
\description{
  Data names of Hills in La Tigra National Park, Honduras
}
\examples{
  if (requireNamespace("sf", quietly = TRUE)) {
    library(sf)
    data(Hills)
    plot(Hills[2], axes=TRUE, pch=16)
  }
}
\references{
  \url{https://www.ign.hn/index.php}
}
\keyword{datasets}
\keyword{sf}
