% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Holdridge.R
\docType{data}
\name{Holdridge}
\alias{Holdridge}
\title{Data for Holdridge Life Zones for La Tigra National Park, Honduras}
\format{
  Simple feature collection with 3 features and 3 fields:
    \itemize{
      \item{id} {}
      \item{LifeZone} {}
      \item{Area_ha} {}
    }
}
\source{
  Leslie Holdridge, 1962. Mapa Ecológico de Honduras.
}
\usage{
  Holdridge
}
\description{
  Data for Holdridge Life Zones for La Tigra National Park, Honduras
}
\examples{
  if (requireNamespace("sf", quietly = TRUE)) {
    library(sf)
    data(Holdridge)
    plot(Holdridge[2], axes=TRUE, col=c("green", "darkgreen", "orange"))
  }
}
\references{
  \url{https://searchworks.stanford.edu/view/2788989}
}
\keyword{datasets}
\keyword{sf}
