% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LULC2020.R
\docType{data}
\name{LULC2020}
\alias{LULC2020}
\title{Data for Land Use and Landcover for year 2020 in La Tigra National Park, Honduras}
\format{
  Simple feature collection with 8 features and 3 fields:
    \itemize{
      \item{id} {}
      \item{LULC} {}
      \item{Area_ha} {}
    }
}
\source{
  Machine learning spatial prediction for land use an landcover using Sentinel-2 Images for 2020.
}
\usage{
  LULC2020
}
\description{
  Data for Land Use and Landcover for year 2020 in La Tigra National Park, Honduras
}
\examples{
  if (requireNamespace("sf", quietly = TRUE)) {
    library(sf)
    data(LULC2020)
    plot(LULC2020[2], axes=TRUE)
  }
}
\references{
  \url{https://www.amitigra.com/}
}
\keyword{datasets}
\keyword{sf}
