% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Microbasins.R
\docType{data}
\name{Microbasins}
\alias{Microbasins}
\title{Data for Microbasins within La Tigra National Park, Honduras}
\format{
    Simple feature collection with 19 features and 3 fields:
    \itemize{
      \item{id} {}
      \item{name} {}
      \item{Area_ha} {}
    }
}
\source{
  Derived from the 30 meters SRTM digital elevation model
}
\usage{
  Microbasins
}
\description{
  Data for Microbasins within La Tigra National Park, Honduras
}
\examples{
  if (requireNamespace("sf", quietly = TRUE)) {
    library(sf)
    data(Microbasins)
    plot(st_geometry(Microbasins), axes=TRUE)
  }
}
\references{
  \url{https://acchonduras.files.wordpress.com/2014/07/pm-pnltfinal.pdf}
}
\keyword{datasets}
\keyword{sf}
