% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Municipalities.R
\docType{data}
\name{Municipalities}
\alias{Municipalities}
\title{Data for Municipalities within and nearby La Tigra National Park, Honduras}
\format{
  Simple feature collection with 4 features and 8 fields:
    \itemize{
      \item{GEOCODIGO} {}
      \item{NOMBRE} {}
      \item{POB_2000} {}
      \item{AREA} {}
      \item{PERIMETER} {}
      \item{HECTARES} {}
      \item{KM2} {}
      \item{Pob_Total} {}
    }
}
\source{
  Instituto Nacional de Estadística, Censo Poblacional de la República de Honduras year 2001
}
\usage{
  Municipalities
}
\description{
  Data for Municipalities within and nearby La Tigra National Park, Honduras
}
\examples{
  if (requireNamespace("sf", quietly = TRUE)) {
    library(sf)
    data(Municipalities)
    plot(st_geometry(Municipalities), axes=TRUE)
  }
}
\references{
  \url{http://170.238.108.227/binhnd/RpWebEngine.exe/Portal?BASE=CPVHND2001&lang=ESP}
}
\keyword{datasets}
\keyword{sf}
