% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rivers.R
\docType{data}
\name{Rivers}
\alias{Rivers}
\title{Data for Rivers for La Tigra National Park, Honduras}
\format{
  Simple feature collection with 196 features and 2 fields:
    \itemize{
      \item{id} {}
      \item{name} {}
    }
}
\source{
  Rivers of Honduras, Instituto Geográfico Nacional, escala 1:50,000.
}
\usage{
  Rivers
}
\description{
  Data for Rivers for La Tigra National Park, Honduras
}
\examples{
  if (requireNamespace("sf", quietly = TRUE)) {
    library(sf)
    data(Rivers)
    plot(Rivers[1], axes=TRUE, col="darkblue")
  }
}
\references{
  \url{https://www.ign.hn/index.php}
}
\keyword{datasets}
\keyword{sf}
