% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Roads.R
\docType{data}
\name{Roads}
\alias{Roads}
\title{Data for Roads within and nearby La Tigra National Park, Honduras}
\format{
  Simple feature collection with 739 features and 5 fields:
    \itemize{
      \item{CODIGO} {}
      \item{TIPO} {}
      \item{Shape_Leng} {}
      \item{id} {}
    }
}
\source{
  Secretaría de Obras Públicas, Transporte y Vivienda de Honduras, 2000 (Hoy INSEP).
}
\usage{
  Roads
}
\description{
  Data for Roads within and nearby La Tigra National Park, Honduras
}
\examples{
  if (requireNamespace("sf", quietly = TRUE)) {
    library(sf)
    data(Roads)
    plot(st_geometry(Roads), axes=TRUE)
  }
}
\references{
  \url{https://www.insep.gob.hn/}
}
\keyword{datasets}
\keyword{sf}
