% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Schools.R
\docType{data}
\name{Schools}
\alias{Schools}
\title{Data for Schools within La Tigra National Park, Honduras}
\format{
    Simple feature collection with 36 features and 17 fields:
    \itemize{
      \item{centroEduc} {}
      \item{coordenada} {}
      \item{coordena_1} {}
      \item{coordena_2} {}
      \item{CE} {}
      \item{Nivel} {}
      \item{Municipio} {}
      \item{Aldea} {}
      \item{Caserio} {}
      \item{Direccion} {}
      \item{tipo_centr} {}
      \item{lug_clases} {}
      \item{nombre_cen} {}
      \item{total_alum} {}
      \item{aulas} {}
      \item{total} {}
      \item{id} {}
    }
}
\source{
  Unidad de Planeamiento y Evaluación la Gestión de la Secretaría de Educación de Honduras, sistema de estadística educativa, 2021.
}
\usage{
  Schools
}
\description{
  Data for Schools within La Tigra National Park, Honduras
}
\examples{
  if (requireNamespace("sf", quietly = TRUE)) {
    library(sf)
    data(Schools)
    plot(st_geometry(Schools), axes=TRUE)
  }
}
\references{
  \url{http://estadisticas.se.gob.hn/see/busqueda.php}
}
\keyword{datasets}
\keyword{sf}
