% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SiteNames.R
\docType{data}
\name{SiteNames}
\alias{SiteNames}
\title{Data names of peculiar sites in La Tigra National Park, Honduras}
\format{
  Simple feature collection with 100 features and 3 fields:
    \itemize{
      \item{id} {}
      \item{name} {}
      \item{order} {}
    }
}
\source{
  Instituto Geográfico Nacional, escala 1:50,000.
}
\usage{
  SiteNames
}
\description{
  Data names of peculiar sites in La Tigra National Park, Honduras
}
\examples{
  if (requireNamespace("sf", quietly = TRUE)) {
    library(sf)
    data(SiteNames)
    plot(SiteNames[2], axes=TRUE, pch=16)
  }
}
\references{
  \url{https://www.ign.hn/index.php}
}
\keyword{datasets}
\keyword{sf}
