% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trails.R
\docType{data}
\name{Trails}
\alias{Trails}
\title{Data of trails for turistic use in La Tigra National Park, Honduras}
\format{
  Simple feature collection with 8 features and 3 fields:
    \itemize{
      \item{names} {}
      \item{X.length_km} {}
      \item{length_m} {}
    }
}
\source{
  Instituto de Conservación Forestal y Vida Silvestre, 2013, Plan de Manejo Parque Nacional La Tigra 2013 - 2025, 219p.
}
\usage{
  Trails
}
\description{
  Data of trails for turistic use in La Tigra National Park, Honduras
}
\examples{
  if (requireNamespace("sf", quietly = TRUE)) {
    library(sf)
    data(Trails)
    plot(st_geometry(Trails), axes=TRUE)
  }
}
\references{
  \url{https://acchonduras.files.wordpress.com/2014/07/pm-pnltfinal.pdf}
}
\keyword{datasets}
\keyword{sf}
