% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Villages.R
\docType{data}
\name{Villages}
\alias{Villages}
\title{Data for Villages within and nearby La Tigra National Park, Honduras}
\format{
  Simple feature collection with 42 features and 6 fields:
    \itemize{
      \item{GEOCODIGO} {}
      \item{AREA} {}
      \item{PERIMETER} {}
      \item{NOMBRE} {}
      \item{POBTOTAL2001} {}
      \item{ID} {}
    }
}
\source{
  Instituto Nacional de Estadística, Censo Poblacional de la República de Honduras year 2001
}
\usage{
  Villages
}
\description{
  Data for Villages within and nearby La Tigra National Park, Honduras
}
\examples{
  if (requireNamespace("sf", quietly = TRUE)) {
    library(sf)
    data(Villages)
    plot(st_geometry(Villages), axes=TRUE)
  }
}
\references{
  \url{http://170.238.108.227/binhnd/RpWebEngine.exe/Portal?BASE=CPVHND2001&lang=ESP}
}
\keyword{datasets}
\keyword{sf}
