% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisitorsCenters.R
\docType{data}
\name{Trails}
\alias{Trails}
\title{Data of Visitors Centers for turistic use in La Tigra National Park, Honduras}
\format{
  Simple feature collection with 2 features and 2 fields:
    \itemize{
      \item{id} {}
      \item{name} {}
    }
}
\source{
  Instituto de Conservación Forestal y Vida Silvestre, 2013, Plan de Manejo Parque Nacional La Tigra 2013 - 2025, 219p.
}
\usage{
  VisitorsCenters
}
\description{
  Data of Visitors Centers for turistic use in La Tigra National Park, Honduras
}
\examples{
  if (requireNamespace("sf", quietly = TRUE)) {
    library(sf)
    data(VisitorsCenters)
    plot(st_geometry(VisitorsCenters), axes=TRUE)
  }
}
\references{
  \url{https://acchonduras.files.wordpress.com/2014/07/pm-pnltfinal.pdf}
}
\keyword{datasets}
\keyword{sf}
