% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WaterIntakeStructure.R
\docType{data}
\name{WaterIntakeStructure}
\alias{WaterIntakeStructure}
\title{Data for Water Intake Structure within La Tigra National Park, Honduras}
\format{
  Simple feature collection with 54 features and 4 fields:
    \itemize{
      \item{Id} {}
      \item{X} {}
      \item{Y} {}
      \item{name_commu} {}
    }
}
\source{
    Instituto de Conservación Forestal y Vida Silvestre, 2013, Plan de Manejo Parque Nacional La Tigra 2013 - 2025, 219p.
}
\usage{
  WaterIntakeStructure
}
\description{
  Data for Water Intake Structures within La Tigra National Park, Honduras
}
\examples{
  if (requireNamespace("sf", quietly = TRUE)) {
    library(sf)
    data(WaterIntakeStructure)
    plot(st_geometry(WaterIntakeStructure), axes=TRUE)
  }
}
\references{
  \url{http://170.238.108.227/binhnd/RpWebEngine.exe/Portal?BASE=CPVHND2001&lang=ESP}
}
\keyword{datasets}
\keyword{sf}
